removed redendant files, added personalize plugin to page-props
