+++ packages/create-sitecore-jss/src/initializers/nextjs-personalize/index.ts
@@ -4,13 +4,8 @@ import {
   openPackageJson,
   transform,
   DEFAULT_APPNAME,
-  // styleguidePrompts,
-  // StyleguideAnswer,
-  // StyleguideArgs,
   ClientAppArgs,
 } from '../../common';
-
-// type NextjsStyleguideArgs = ClientAppArgs & StyleguideArgs;
 
 export default class NextjsPersonalizeInitializer implements Initializer {
   get isBase(): boolean {
@@ -20,14 +15,13 @@ export default class NextjsPersonalizeInitializer implements Initializer {
   async init(args: ClientAppArgs) {
     const pkg = openPackageJson(`${args.destination}${sep}package.json`);
 
-    // TODO: prompts for Personalize
+    // TODO: prompts for Personalize and argument types
     // const answers = await prompt<StyleguideAnswer>(styleguidePrompts, args);
 
     const mergedArgs = {
       ...args,
       appName: args.appName || pkg?.config?.appName || DEFAULT_APPNAME,
       appPrefix: args.appPrefix || pkg?.config?.prefix || false,
-      // ...answers,
     };
 
     const templatePath = path.resolve(__dirname, '../../templates/nextjs-personalize');
