+++ packages/create-sitecore-jss/src/templates/nextjs-personalize/.env
@@ -1,4 +1,3 @@
 BOXEVER_CLIENT_KEY=
 BOXEVER_API=
 BOXEVER_TARGET_URL=
-GRAPH_QL_ENDPOINT=
