--- packages/create-sitecore-jss/src/templates/nextjs-personalize/data/dictionary/da-DK.yml
@@ -1,4 +1,0 @@
-styleguide-sample: Dette er en ordbogspost på dansk som en demonstration
-Documentation: Dokumentation
-Styleguide: Styleguide
-GraphQL: GraphQL
