--- packages/create-sitecore-jss/src/templates/nextjs-personalize/data/dictionary/en.yml
@@ -1,4 +1,0 @@
-styleguide-sample: This is a dictionary entry in English as a demonstration
-Documentation: Documentation
-Styleguide: Styleguide
-GraphQL: GraphQL
