--- packages/create-sitecore-jss/src/templates/nextjs-personalize/data/routes/da-DK.yml
@@ -1,4 +1,0 @@
-# in order to demonstrate translation on the Styleguide page, we need to have a stub on the homepage for that language.
-# this can be deleted safely if not using multiple languages - or not using Danish.
-name: home
-id: home-page
