--- packages/create-sitecore-jss/src/templates/nextjs-personalize/data/routes/graphql/sample-1/en.yml
@@ -1,9 +1,0 @@
-fields:
-  pageTitle: Sample 1 Page Title
-placeholders:
-  jss-main:
-  - componentName: ContentBlock
-    fields:
-      heading: GraphQL Sample 1
-      content: |
-        <p>A child route here to illustrate the power of GraphQL queries. <a href="/graphql">Back to GraphQL route</a></p>
