--- packages/create-sitecore-jss/src/templates/nextjs-personalize/src/lib/component-props.ts
@@ -1,6 +1,0 @@
-﻿import { ComponentRendering} from '@sitecore-jss/sitecore-jss-nextjs';
-
-// NULL means Hidden by this experience
-export type ComponentRenderingWithExpiriences = ComponentRendering & {
-  experiences: { [name: string]: ComponentRenderingWithExpiriences | null };
-};
