+++ packages/create-sitecore-jss/src/templates/nextjs/src/lib/component-props/index.ts
@@ -21,3 +21,8 @@ export type ComponentProps = {
 export type ComponentWithContextProps = ComponentProps & {
   sitecoreContext: SitecoreContextValue;
 };
+
+// NULL means Hidden by this experience
+export type ComponentRenderingWithExpiriences = ComponentRendering & {
+  experiences: { [name: string]: ComponentRenderingWithExpiriences | null };
+};
